UNIQPTR_T moduled
