+++ include/bk/bk.h
@@ -11,7 +11,8 @@
 
 #include <string_view>
 #include <vector>
-#include "addr.h"
+#include "uniptr.h"
+#include "bk/addr.h"
 
 class TX_T;
 class BK_T {
