+++ include/common.h
@@ -20,28 +20,6 @@ inline std::string_view u8string_view(const u8_t *s, size_t size) {
 inline std::string_view u256string_view(const uint256_t &key) {
   return u8string_view(std::data(key), sizeof(uint256_t));
 }
-/// Universal ptr
-union   UNIPTR_T {
-    const void        *v_ptr;
-    const char        *ch_ptr;
-    const u8_t        *u8_ptr;
-    const uint16_t    *u16_ptr;
-    const uint32_t    *u32_ptr;
-    const uint64_t    *u64_ptr;
-    const uint160_t   *u160_ptr;
-    const uint256_t   *u256_ptr;
-    UNIPTR_T() {}
-    UNIPTR_T(const char * ptr) { ch_ptr = ptr; }
-    UNIPTR_T(const uint160_t * ptr) { u160_ptr = ptr; }
-    UNIPTR_T(const uint256_t * ptr) { u256_ptr = ptr; }
-    uint32_t          take_32(void) { return *u32_ptr++; }
-    uint64_t          take_64(void) { return *u64_ptr++; }
-    uint32_t          take_varuint(void);
-    const u8_t        *take_u8_ptr(const uint32_t);
-    const char        *take_ch_ptr(const uint32_t);
-    const uint32_t    *take_32_ptr(void) { return u32_ptr++; }
-    const uint256_t   *take_256_ptr(void) { return u256_ptr++; }
-};
 
 class BCException : public std::runtime_error {
 public:
